Put cncserver client in separate file
